Add persisted connection parameters to connection state [VPNLINUX-436]
When a connection was previously persisted to disk, the VPNConnection
instance is initialized with the Connected state.

This commit adds the persisted connection parameters to the Connected
state context, so that subscribers to connection state changes can
access these parameters (e.g. to know the name/id of the VPN server
the connection was established to).