Check TCP connectivity to the server before starting a VPN connection [VPNLINUX-568]
The reason for doing this is that, since we added the IPv6 leak
protection implemented with a dummy network interface, we've been
having issues reconnecting to the VPN server after the machine
wakes up from suspend.

The root cause of these issues was that the dummy interface to
protect against IPv6 leaks was being used to establish the VPN
connection, since when the VPN goes down it's the interface
with the highest priority.

To avoid the issue, before starting the VPN connection network
connectivity is probed by trying to open a socket to any of
the TCP ports the VPN server is listening to.

The reason why we do this connection check using TCP ports but
not UDP ones is only that checking if a UDP port is open not
as trivial, and checking for TCP only might be good enough.