Add delay before making tcpcheck call [VPNLINUX-620]
This delays is needed mainly when using NetworkManager as a kill switch. Since it takes some time after a dummy connection is added to NM and it being active, if in the meantime the tcpcheck is scheduled, it will possibly only block the call, making it look like the server is unreacheable.