Initialize state machine in Error state when inactive VPN connection is found [VPNLINUX-747]
If an inactive VPN connection is found when initializing the state
machine then it means that previously a VPN connection was established but
it dropped unexpectedly.

Note that when the user willingly disconnects from the VPN, then the NM
connection is discarded so no inactive connections are left over in this
scenario.

Initializing the state machine in error state when an inactive VPN connection
is critical so that the kill switch behaves properly. The error state does not
disable the kill switch, therefore it prevents leaks. However, in disconnected
state, the KS is automatically disabled (unless configured as "permanent").