[VPNLINUX-968] Fix app stuck in disconnecting state
When the app ends up in error state, the underlying network manager
connection is often removed automatically by network manager. When
that happens, if the user or the reconnector tries to manually
disconnect, a Disconnected event is never sent because there was
no underlying connection in the first place.

In these cases, we now emit an artificial Disconnected event, so that
the connection state transitions from Error to Disconnecting and finally
to Disconnected (after it receives the Disconnected event).