[VPNLINUX-968] Fix app stuck in disconnecting state on OpenVPN
When the app ends up in error state, the underlying OpenVPN
network manager connection is not active anymore. If the app
tries to stop the connection, it hangs waiting to receive the
Disconnected state, which it never receives because the NM
connection was not active in the first place.

In these cases, we now emit an artificial Disconnected event, so that
the connection state transitions from Error to Disconnecting and finally
to Disconnected (after it receives the Disconnected event).